Added DelayPatch operation.
