+++ Source/Control.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Xml;
+﻿using System.Xml;
 using Verse;
 
 namespace XmlExtensions
@@ -111,384 +110,26 @@ namespace XmlExtensions
 
     }
 
-    public class And : PatchOperationBoolean
+    public class DelayPatches : PatchOperation
     {
-        protected PatchOperationBoolean condition1 = null;
-        protected PatchOperationBoolean condition2 = null;
+        protected XmlContainer patches = null;
+        protected bool applyHere = false;
 
         protected override bool ApplyWorker(XmlDocument xml)
         {
-            return true;
-        }
-        protected override bool evaluation(XmlDocument xml)
-        {
-            return this.condition1.evaluate(xml) && this.condition2.evaluate(xml);
-        }
-    }
-
-    public class Or : PatchOperationBoolean
-    {
-        protected PatchOperationBoolean condition1 = null;
-        protected PatchOperationBoolean condition2 = null;
-
-        protected override bool ApplyWorker(XmlDocument xml)
-        {
-            return true;
-        }
-        protected override bool evaluation(XmlDocument xml)
-        {
-            return this.condition1.evaluate(xml) || this.condition2.evaluate(xml);
-        }
-    }
-
-    public class Not : PatchOperationBoolean
-    {
-        protected PatchOperationBoolean condition = null;
-
-        protected override bool ApplyWorker(XmlDocument xml)
-        {
-            return true;
-        }
-        protected override bool evaluation(XmlDocument xml)
-        {
-            return !this.condition.evaluate(xml);
-        }
-    }
-
-    public class Comparision : PatchOperationBoolean
-    {
-        protected string value = "0";
-        protected string relation = "=";
-        protected string logic = "or";
-        protected string fromXml = "false";
-        protected string isFloat = "false";
-
-        protected override bool ApplyWorker(XmlDocument xml)
-        {
-            return true;
-        }
-
-        protected override bool evaluation(XmlDocument xml)
-        {
-            bool isOr = true;
-            if (this.logic == "and")
+            for (int i = 0; i < this.patches.node.ChildNodes.Count; i++)
             {
-                isOr = false;
-            }
-            flag = !isOr;
-            if (!bool.Parse(this.isFloat))
-            {
-                string val2 = "";
-                if (fromXml == "true")
+                XmlNode node = patches.node.ChildNodes[i];
+                PatchManager.enqueuePatch(node);
+                if (applyHere)
                 {
-                    val2 = xml.SelectSingleNode(value).InnerText;
-                }
-                else
-                {
-                    val2 = value;
-                }
-
-                foreach (object obj in xml.SelectNodes(this.xpath))
-                {
-                    XmlNode xmlNode = obj as XmlNode;
-                    string xval = xmlNode.InnerText;
-                    int compare = xval.CompareTo(val2);
-                    if (relation == "eq")
-                    {
-                        if (isOr)
-                        {
-                            if (compare == 0)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (compare == 0)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "sl")
-                    {
-                        if (isOr)
-                        {
-                            if (compare < 0)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (compare < 0)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "leq")
-                    {
-                        if (isOr)
-                        {
-                            if (compare <= 0)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (compare <= 0)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "sg")
-                    {
-                        if (isOr)
-                        {
-                            if (compare > 0)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (compare > 0)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "geq")
-                    {
-                        if (isOr)
-                        {
-                            if (compare >= 0)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (compare >= 0)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "neq")
-                    {
-                        if (isOr)
-                        {
-                            if (compare != 0)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (compare != 0)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else
-                    {
-                        return false;
-                    }
+                    PatchOperation patch = Helpers.getPatchFromString(node.OuterXml);
+                    patch.Apply(xml);
                 }
             }
-            else
-            {
-                float val2 = 0;
-                if (fromXml == "true")
-                {
-                    val2 = float.Parse(xml.SelectSingleNode(value).InnerText);
-                }
-                else
-                {
-                    val2 = float.Parse(value);
-                }
-                foreach (object obj in xml.SelectNodes(this.xpath))
-                {
-                    XmlNode xmlNode = obj as XmlNode;
-                    float xval = float.Parse(xmlNode.InnerText);
-                    if (relation == "eq")
-                    {
-                        if (isOr)
-                        {
-                            if (xval == val2)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (xval == val2)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "sl")
-                    {
-                        if (isOr)
-                        {
-                            if (xval < val2)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (xval < val2)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "leq")
-                    {
-                        if (isOr)
-                        {
-                            if (xval <= val2)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (xval <= val2)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "sg")
-                    {
-                        if (isOr)
-                        {
-                            if (xval > val2)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (xval > val2)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "geq")
-                    {
-                        if (isOr)
-                        {
-                            if (xval >= val2)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (xval >= val2)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else if (relation == "neq")
-                    {
-                        if (isOr)
-                        {
-                            if (xval != val2)
-                            {
-                                flag = true;
-                                return flag;
-                            }
-                        }
-                        else
-                        {
-                            if (xval != val2)
-                            {
-                            }
-                            else
-                            {
-                                flag = false;
-                                return flag;
-                            }
-                        }
-                    }
-                    else
-                    {
-                        return false;
-                    }
-                }
-            }
-
-                return flag;
+            return true;
         }
-        }
-
+    }
     
 }
