+++ Source/Patch_ApplyPatches.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections.Generic;
+using System.Xml;
+using HarmonyLib;
+using Verse;
+
+namespace XmlExtensions
+{
+
+    [HarmonyPatch(typeof(LoadedModManager))]
+    [HarmonyPatch("ApplyPatches")]
+    class Patch_ApplyPatches
+    {
+        static void Postfix(XmlDocument xmlDoc, Dictionary<XmlNode, LoadableXmlAsset> assetlookup)
+        {
+            PatchManager.applyPatches(xmlDoc);
+        }
+    }
+}
