+++ Source/XmlExtensions.csproj
@@ -21,6 +21,9 @@
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="0Harmony">
+      <HintPath>..\..\..\..\..\..\..\..\Users\15adh\Desktop\0Harmony.dll</HintPath>
+    </Reference>
     <Reference Include="Assembly-CSharp">
       <HintPath>..\..\..\RimWorldWin64_Data\Managed\Assembly-CSharp.dll</HintPath>
       <Private>False</Private>
@@ -45,7 +48,10 @@
     <Compile Include="Data.cs" />
     <Compile Include="Helpers.cs" />
     <Compile Include="Mod_Settings.cs" />
+    <Compile Include="PatchManager.cs" />
+    <Compile Include="PatchOperationBoolean.cs" />
     <Compile Include="PatchOperations.cs" />
+    <Compile Include="Patch_ApplyPatches.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="XmlExtensions.cs" />
   </ItemGroup>
